style(makefile): split calm.so main.exe
