--- CalmClam.h
@@ -1,9 +1,0 @@
-#ifndef CALMCLAM_H
-#define CALMCLAM_H
-
-#include "myStruct.h"
-#include "Kitchen.h"
-#include "Chef.h"
-#include "Func.h"
-
-#endif /* CALMCLAM_H */
