+++ README.md
@@ -1,18 +1,18 @@
 # CALM
-coding game code a la mode 實作
 
-# Include
-```c=
+An simple library contain some useful utility to beat `coding game code a la`.
+
+## Requirement
+
+- [GCC] 4.7 above (support C++11)
+
+## Usage
+
+```c++
 #include "CalmClam.h"
 ```
 
-# Complie
-g++ Kitchen.cpp Chef.cpp Func.cpp main.cpp -o {output filename}
-or
-make
+## Document
 
-# Document
-[link](https://hackmd.io/@TPPG2/HyvEG8NXc)
-
-# Merge into one file
-https://www.npmjs.com/package/codingame-cpp-merge
+- [wiki](https://hackmd.io/@TPPG2/HyvEG8NXc)
+- more info: https://www.npmjs.com/package/codingame-cpp-merge
